Updated Login page
Pending logic implementation
